add aic_mode and easy_enable cmds

+++ src/command/set_sbas_enabled.rs
@@ -14,7 +14,7 @@ impl Command for SetSbasEnabled {
     type Response = Ack;
 
     fn encode(&self) -> Result<PmtkPacket, PmtkError> {
-        let data_field = encode_data_field([self.0 as u32]);
+        let data_field = encode_data_field([self.0 as u8]);
         PmtkPacket::new_command(Self::PKT_TYPE, Some(data_field))
     }
 }
